Move get_ca_certs_path to datadog_checks_base as it's shared by both http_check and tcp_check
