+++ datadog_checks_base/datadog_checks/utils/ca_cert.py
@@ -1,0 +1,4 @@
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from ..base.utils.ca_cert import _get_ca_certs_paths, get_ca_certs_path
