+++ datadog_checks_base/tests/test_utils.py
@@ -1,9 +1,15 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import os
+import sys
+import pytest
+import mock
 
 from datadog_checks.utils.common import pattern_filter
 from datadog_checks.utils.limiter import Limiter
+from datadog_checks.utils.ca_cert import _get_ca_certs_paths, get_ca_certs_path
+from datadog_checks.dev import temp_dir
 
 
 class Item:
@@ -105,3 +111,49 @@ class TestLimiter():
         assert limiter.get_status() == (0, 10, False)
         assert limiter.is_reached("dummy1") is False
         assert limiter.get_status() == (1, 10, False)
+
+
+@pytest.mark.unit
+def test_get_ca_certs_path():
+    with mock.patch('datadog_checks.base.utils.ca_cert._get_ca_certs_paths') as gp:
+        # no certs found
+        gp.return_value = []
+        assert get_ca_certs_path() is None
+        # one cert file was found
+        # let's avoid creating a real file just for the sake of mocking a cert
+        # and use __file__ instead
+        gp.return_value = [__file__]
+        assert get_ca_certs_path() == __file__
+
+
+@pytest.mark.unit
+def test__get_ca_certs_paths_ko():
+    """
+    When `embedded` folder is not found, it should raise OSError
+    """
+    with pytest.raises(OSError):
+        _get_ca_certs_paths()
+
+
+@pytest.mark.unit
+def test__get_ca_certs_paths():
+    with mock.patch('datadog_checks.base.utils.ca_cert.os.path.dirname') as dirname:
+        # create a tmp `embedded` folder
+        with temp_dir() as tmp:
+            target = os.path.join(tmp, 'embedded')
+            os.mkdir(target)
+            # point `dirname()` there
+            dirname.return_value = target
+
+            # tornado not found
+            paths = _get_ca_certs_paths()
+            assert len(paths) == 2
+            assert paths[0].startswith(target)
+            assert paths[1] == '/etc/ssl/certs/ca-certificates.crt'
+
+            # mock tornado's presence
+            sys.modules['tornado'] = mock.MagicMock(__file__='.')
+            paths = _get_ca_certs_paths()
+            assert len(paths) == 3
+            assert paths[1].endswith('ca-certificates.crt')
+            assert paths[2] == '/etc/ssl/certs/ca-certificates.crt'
